Fix: Task-I3---003--Developer Dashboard

+++ src/main/java/acme/forms/DeveloperDashboard.java
@@ -15,13 +15,13 @@ public class DeveloperDashboard extends AbstractForm {
 
 	// Attributes -------------------------------------------------------------
 
-	Double						totalNumberOfTrainingModulesWithUpdatedMoment;
+	Integer						totalNumberOfTrainingModulesWithUpdatedMoment;
 
-	Double						totalNumberOfTrainingSessionsWithLink;
+	Integer						totalNumberOfTrainingSessionsWithLink;
 
 	Double						averageTimeTrainingModules;
 	Double						deviationTimeTrainingModules;
-	Double						minimumTimeTrainingModules;
-	Double						maximumTimeTrainingModules;
+	Integer						minimumTimeTrainingModules;
+	Integer						maximumTimeTrainingModules;
 
 }
